+++ src/core/cache/default/spec.js
@@ -6,23 +6,43 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
-import SimpleCache from 'core/cache/simple';
+import DefaultCache from 'core/cache/default';
 
 describe('core/cache/simple', () => {
+	it('default value', () => {
+		const
+			cache = new DefaultCache(Array);
+
+		expect(cache.has('foo')).toBeFalse();
+
+		expect(cache.get('foo')).toEqual([]);
+		expect(cache.size).toBe(1);
+
+		cache.defaultFactory = () => 10;
+
+		expect(cache.get('bla')).toEqual(10);
+		expect(cache.size).toBe(2);
+
+		expect(cache.set('bar', 10)).toBe(10);
+		expect(cache.size).toBe(3);
+	});
+
 	it('crud', () => {
-		const cache = new SimpleCache();
-		expect(cache.has('foo')).toBe(false);
+		const
+			cache = new DefaultCache();
+
+		expect(cache.has('foo')).toBeFalse();
 		expect(cache.set('foo', 1)).toBe(1);
 		expect(cache.get('foo')).toBe(1);
-		expect(cache.has('foo')).toBe(true);
+		expect(cache.has('foo')).toBeTrue();
 		expect(cache.size).toBe(1);
 		expect(cache.remove('foo')).toBe(1);
-		expect(cache.has('foo')).toBe(false);
+		expect(cache.has('foo')).toBeFalse();
 	});
 
 	it('default iterator', () => {
 		const
-			cache = new SimpleCache();
+			cache = new DefaultCache();
 
 		cache.set('1', 1);
 		cache.set('2', 2);
@@ -33,7 +53,7 @@ describe('core/cache/simple', () => {
 
 	it('`keys`', () => {
 		const
-			cache = new SimpleCache();
+			cache = new DefaultCache();
 
 		cache.set('1', 1);
 		cache.set('2', 2);
@@ -43,7 +63,7 @@ describe('core/cache/simple', () => {
 
 	it('`values`', () => {
 		const
-			cache = new SimpleCache();
+			cache = new DefaultCache();
 
 		cache.set('1', 1);
 		cache.set('2', 2);
@@ -53,7 +73,7 @@ describe('core/cache/simple', () => {
 
 	it('`entries`', () => {
 		const
-			cache = new SimpleCache();
+			cache = new DefaultCache();
 
 		cache.set('1', 1);
 		cache.set('2', 2);
@@ -62,32 +82,34 @@ describe('core/cache/simple', () => {
 	});
 
 	it('`clear`', () => {
-		const cache = new SimpleCache();
+		const
+			cache = new DefaultCache();
 
 		cache.set('foo', 1);
 		cache.set('bar', 2);
 
-		expect(cache.has('foo')).toBe(true);
-		expect(cache.has('bar')).toBe(true);
+		expect(cache.has('foo')).toBeTrue();
+		expect(cache.has('bar')).toBeTrue();
 
 		expect(cache.clear()).toEqual(new Map([['foo', 1], ['bar', 2]]));
 	});
 
 	it('`clear` with a filter', () => {
-		const cache = new SimpleCache();
+		const
+			cache = new DefaultCache();
 
 		cache.set('foo', 1);
 		cache.set('bar', 2);
 
-		expect(cache.has('foo')).toBe(true);
-		expect(cache.has('bar')).toBe(true);
+		expect(cache.has('foo')).toBeTrue();
+		expect(cache.has('bar')).toBeTrue();
 
 		expect(cache.clear((el) => el > 1)).toEqual(new Map([['bar', 2]]));
 	});
 
 	it('`clones`', () => {
 		const
-			cache = new SimpleCache(),
+			cache = new DefaultCache(),
 			obj = {a: 1};
 
 		cache.set('foo', 1);
@@ -102,6 +124,6 @@ describe('core/cache/simple', () => {
 		expect(newCache.has('foo')).toBeTrue();
 		expect(newCache.has('bar')).toBeTrue();
 		expect(cache.storage !== newCache.storage).toBeTrue();
-		expect(cache.get('bar') === newCache.get('bar')).toBeTrue();
+		expect(cache.get('bla') === newCache.get('bla')).toBeTrue();
 	});
 });